Extract color_picker and file_uploader
